feat(reveal): add reveal strategies
This adds different reveal strategies to the 'reveal' method.
Reveal strategies can be one of the following:
'left' | 'right' | 'inside' | 'outside' | 'random', where 'random' is the default.
BREAKING CHANGE: In order for the reveal method to not have an exhaustive list of params,
it introduces an object as a parameter. This makes the api much clearer.
Furthermore, it introduces a dependency called 'is-odd'
(https://www.npmjs.com/package/is-odd) that checks whether
a given number is odd (needed for 'outside reveal strategy').

Fixes #32